Tried to move ElementCtx outside of Elements.
Tried to move ElementCtx outside of the definition of Elements, removing ctx() and ctx_mut() from the trait. Element + Ctx are wrapped in an ElemBox. Trait methods take an ElementMut which is a RcBorrow but with mutable access.

This failed because ElementMut doesn't automatically reborrow like regular `&mut T` references, and this _greatly_ impacts the ergonomics. Unfortunately there doesn't seem to be any plan in the language to specify reborrowing semantics for reference-like types.