[kyute] Clip CompositionBuilder layers to the parent window area; round up text measure to the nearest integer; round window size to nearest larger integer size
Previously, fractional window sizes were truncated (by winit), resulting in windows that were too small to hold the content.
With text elements, this resulted in a relayout that made the content overflow the window bounds.
In this case, CompositionBuilder allocated another layer for the overflowing content, which is not what we want (we can't paint outside the window bounds, even on a separate layer).

This commit adds guardrails to prevent this:
- initial window sizes are rounded up to the next logical pixel
- picture layers implicitly created by CompositionBuilder are clipped to the window area
- `Text::measure` rounds up the size to the next logical pixel