[kyute] remove SizeConstraint
SizeConstraint::Unspecified was used to query the ideal size of an element, but it was essentially equivalent to measuring with an infinite available space. So remove SizeConstraint and just pass a `Size` value directly.

This doesn't seem to break anything in fluff or gallery.
But then again there are no tests, and no tests => no problems :)